Moving component NewDao in a new file
